Authentication, tags

+++ resources/views/posts/show.blade.php
@@ -3,12 +3,33 @@
 @section('content')
 
     <b>Title:</b>
+
     <p>{{ $post->title }}</p>
+
     <b>Text:</b>
+
     <p>{!! nl2br($post->body) !!}</p>
+
+    @unless($post->tags->isEmpty())
+
+        <b>Tags:</b>
+
+        <ul class="list-unstyled">
+
+            @foreach($post->tags as $tag)
+
+                <li>{!! link_to_route('tags.show', $tag->name, [$tag->id]) !!}</li>
+
+            @endforeach
+
+        </ul>
+
+    @endunless
+
     <p><i>Author: {{ $post->user->name }}</i></p>
 
-    @if($post->user_id==\Illuminate\Support\Facades\Auth::user()->id)
+
+    @if(\Illuminate\Support\Facades\Auth::check() and $post->user_id==\Illuminate\Support\Facades\Auth::user()->id)
 
         {!! link_to_route('posts.edit', 'Edit', [$post->id], ['class' => 'btn btn-primary']) !!}
 
@@ -29,9 +50,9 @@
 
                 <p>{{ $comment->body }}</p>
 
-                <p class="text-muted">{{ $comment->created_at }}</p>
+                <p class="text-muted">{{ $comment->created_at->diffForHumans() }}</p>
 
-                @if($comment->user_id==\Illuminate\Support\Facades\Auth::user()->id)
+                @if(\Illuminate\Support\Facades\Auth::check() and $comment->user_id==\Illuminate\Support\Facades\Auth::user()->id)
 
                     {!! delete_comments_form(['comments.destroy', $comment->id], $post->id) !!}
 
@@ -45,8 +66,10 @@
 
     </ul>
 
-    @include('comments.create')
+    @if(\Illuminate\Support\Facades\Auth::check())
 
+        @include('comments.create')
 
+    @endif
 
 @endsection
